pyproject.toml
[project]
name = "python-crawler"
version = "0.1.0"
description = "Web crawler with Crawl4AI, Playwright, Streamlit, ChromaDB, Ollama"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "crawl4ai",
    "playwright",
    "streamlit",
    "chromadb",
    "ddgs",
    "langchain-text-splitters",
    "langchain-community",
    "ollama",
    "unstructured>=0.18.14",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
]

[tool.uv]
# ให้ uv ใช้ python 3.12 แทน 3.13
python-preference = "system"   # หรือ "managed" ถ้าอยากให้ uv จัดการ python runtime ให้
